pyproject.toml
[tool.poetry]
name = "practice-django"
version = "0.1.0"
description = ""
authors = ["Sergei Mikurov <454004@mail.ru>"]
readme = "README.md"
packages = [{include = "practice_django"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "4.2.8"
python-dotenv = "^1.0.0"
django-cleanup = "^8.0.0"
django-ckeditor = "^6.7.0"
channels = "^4.0.0"
django-allauth = "^0.59.0"
django-braces = "^1.15.0"
pillow = "^10.1.0"
django-crispy-forms = "^2.1"
crispy-bootstrap5 = "^2023.10"
django-debug-toolbar = "^4.2.0"
django-taggit = "^5.0.1"
django-bootstrap5 = "^23.3"
python-slugify = "^8.0.1"
pytils = "^0.4.1"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.1.0"
django-extensions = "^3.2.3"
ipython = "^8.18.1"
notebook = "6.5.6"
jupyterthemes = "^0.20.0"
jupyter-contrib-nbextensions = "^0.7.0"
jupyter-nbextensions-configurator = "^0.6.3"



[tool.poetry.group.doc.dependencies]
mkdocs = "^1.5.3"
mkdocs-material = "^9.5.2"
mkdocs-awesome-pages-plugin = "^2.9.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
